+++ Sources/Base/OAuth2Base.swift
@@ -166,10 +166,10 @@ open class OAuth2Base: OAuth2Securable {
 	
 	- keychain (bool, true by default, applies to using the system keychain)
 	- keychain_access_mode (string, value for keychain kSecAttrAccessible attribute, kSecAttrAccessibleWhenUnlocked by default)
+	- keychain_access_group (string, value for keychain kSecAttrAccessGroup attribute, nil by default)
 	- verbose (bool, false by default, applies to client logging)
 	- secret_in_body (bool, false by default, forces the flow to use the request body for the client secret)
 	- token_assume_unexpired (bool, true by default, whether to use access tokens that do not come with an "expires_in" parameter)
-	- title (string, to be shown in views shown by the framework)
 	*/
 	override public init(settings: OAuth2JSON) {
 		clientConfig = OAuth2ClientConfig(settings: settings)