+++ CHANGELOG.md
@@ -11,6 +11,7 @@ You can also refer to commit logs to get details on what was implemented, fixed 
 - Broad API redesign, you should now use `authorize(params:callback:)`
 - All errors returned by OAuth2 are now `OAuth2Error` types
 - Add `Package.swift` for the Swift package manager
+- Expose `keychainAccessGroup` (`keychain_access_group` in settings; thanks @damienrambout !)
 
 
 ### 2.2.9